JKA-1727 ①Container / Presentational 分離 JKA-1722（React 講座詳細ページでAPIからデータを取得する）のサブタスク
